requirements.txt
certifi==2018.1.18
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
future==0.16.0
idna==2.6
pycparser==2.19
python-telegram-bot==12.0.0
requests==2.22.0
six==1.13.0
tornado==6.0.3
urllib3==1.25.7
